# ENH: Imviz parser optimizations