--- a/CHANGES.rst
+++ b/CHANGES.rst
@@ -97,8 +97,10 @@ Cubeviz
 Imviz
 ^^^^^
 
-* Do not hide previous results in aperture photometry when there is a failure, but rather show
+- Do not hide previous results in aperture photometry when there is a failure, but rather show
   the failure message within the plugin UI to indicate the shown results are "out of date". [#2112]
+
+- More efficient parser for Roman data products in Imviz [#2176]
 
 Mosviz
 ^^^^^^