@bmorris3 2023-05-03
One more follow-up. I made a [demo notebook](https://gist.github.com/bmorris3/1df881a5bbde999a925e38385ae6d9bf) to test how efficiently `ImageModels` from `roman_datamodels` are loaded into Imviz, assuming they're already loaded from disk, and in memory. Even with 18 detectors with images of shape (4088, 4088), the elapsed time for making an Imviz instance and loading all detectors with WCS linking is finished in just under 2 seconds. So for now the bottleneck is definitely in loading ASDF files from disk in `roman_datamodels`.